requirements.txt
streamlit
google-generativeai
pydantic
typing_extensions
